Feat: image resource endpoint 추가
