+++ src/files/files.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { FilesController } from './files.controller';
+
+@Module({
+  controllers: [FilesController],
+})
+export class FilesModule {}
