+++ src/users/users.controller.ts
@@ -1,5 +1,4 @@
 import {
-  BadRequestException,
   Body,
   Controller,
   Get,
@@ -14,7 +13,6 @@ import {
   UseInterceptors,
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
-import { MulterOptions } from '@nestjs/platform-express/multer/interfaces/multer-options.interface';
 import {
   ApiConsumes,
   ApiCookieAuth,
@@ -22,33 +20,13 @@ import {
   ApiResponse,
   ApiTags,
 } from '@nestjs/swagger';
-import { diskStorage } from 'multer';
 import { AuthenticatedGuard } from 'src/auth/guards/authenticated.guard';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { GetUser } from './get-user.decorator';
 import { User } from './user.entity';
 import { UsersService } from './users.service';
-
-const localOptions: MulterOptions = {
-  storage: diskStorage({
-    destination: './files/avatar',
-    filename: (req, file, cb) => {
-      console.log(req, file, cb);
-      const fileName: string = req.body.name;
-      const timeStamp = Date.now();
-      const fileExt: string = file.mimetype.split('/')[1];
-      return cb(null, `${fileName}.${timeStamp}.${fileExt}`);
-    },
-  }),
-  fileFilter: (req, file, cb) => {
-    const type: string = file.mimetype.split('/')[0];
-    if (type !== 'image') {
-      return cb(new BadRequestException(`not a image file!`), false);
-    }
-    return cb(null, true);
-  },
-};
+import { localOptions } from './constants';
 
 @ApiTags('Users')
 @Controller('users')
